Add files via upload
I Don't know why but it didn't show the css affects on the html website. You can see I coded the css thing I'm not sure why it didn't work but even if I end up losing marks can I not get penalized for not having it at all because I did make the css file. If I lose marks that's okay as it Is my fault for not fixing the issue.